move get_tournament_round_group_details to own file
