+++ chessdotcom/endpoints/tournament_round_group_details.py
@@ -1,0 +1,22 @@
+from ..client import Client, Resource
+from ..response_builder import ChessDotComResponse
+
+
+@Client.endpoint
+def get_tournament_round_group_details(
+    url_id: str, round_num: int, group_num: int, tts=0, **request_options
+) -> ChessDotComResponse:
+    """
+    :param url_id: URL for the club's web page on www.chess.com.
+    :param round_num: the round of the tournament.
+    :param group_num: the group in the tournament.
+    :param tts: the time the client will wait before making the first request.
+    :returns: :obj:`ChessDotComResponse` object containing
+                details about a tournament's round group.
+    """
+    return Resource(
+        uri=f"/tournament/{url_id}/{round_num}/{group_num}",
+        tts=tts,
+        top_level_attribute="tournament_round_group",
+        request_options=request_options,
+    )
